ecommerce/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==5.1.1
Django==4.0.6
django-rest-framework==0.1.0
djangorestframework==3.13.1
executing==0.10.0
factory-boy==3.2.1
Faker==13.15.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==6.2.5
pytest-base-url==1.4.2
pytest-django==4.5.2
pytest-factoryboy==2.5.0
pytest-html==3.1.1
pytest-metadata==2.0.2
pytest-selenium==3.0.0
pytest-variables==1.9.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
